// src/interfaces.ts
export interface DataModelConstraintSchema {
    /**
     * A string which represents the type of this constraint e.g. unique
     */
    type: "unique";
    /**
     * A short description for this constraint e.g. Unique identifier field must be unique across different records.
     */
    description?: string;
    fields?: string[];
}

export interface DataModelEventListenerSchema {
    /**
     * A string which the name of this event listener e.g. update person user listener
     */
    name?: string;
    /**
     * A string which represents the path of the module that exports this listener. This path may be a relative to execution folder path of a module exists in package modules e.g. ./listeners/add-user-listener or my-module/send-mail-listener
     */
    type: string;
  }

export interface DataModelPrivilegeSchema {
    /**
     * A number which represents permission mask (1=Read, 2=Create, 4=Update, 8=Delete, 16=Execute)
     */
    mask: number;
    /**
     * A string which represents the permission scope.
     */
    type: "self" | "global" | "parent" | "item";
    /**
     * A string which represents the name of the security group where this privilege will be applied e.g. Administrators, Sales etc.
     */
    account?: string;
    /**
     * A string which represents a filter expression for this privilege. This attribute is used for self privileges which are commonly derived from user's attributes e.g. 'owner eq me()' or 'orderStatus eq 1 and customer eq me()' etc.
     */
    filter?: string;
    [k: string]: unknown;
}

export interface DataModelSchema {
  /**
   * A string which represents a literal unique identifier for this model e.g. https://example.com/models/User
   */
  "@id"?: string;
  /**
   * A string which represents the name of this model e.g. Order, Customer, Person etc
   */
  name: string;
  /**
   * An optional numeric identifier for this model e.g. 9587891
   */
  id?: number;
  /**
   * A string which represents the title of this e.g. Supplier Orders, Person Followers etc
   */
  title?: string;
  /**
   * A string which represents the model which is inherited by this model e.g. User inherits Account, Person inherits Party etc
   */
  inherits?: string;
  /**
   * A string which represents the model which is implemented by this model e.g. ActionStatusType model implements Enumeration model etc
   */
  implements?: string;
  /**
   * A boolean which indicates whether this model is being upgraded automatically or not. The default value is false.
   */
  sealed?: boolean;
  /**
   * A boolean which indicates whether this model is an abstract model or not. The default value is false.
   */
  abstract?: boolean;
  /**
   * A boolean which indicates whether this model is hidden or not. The default value is false.
   */
  hidden?: boolean;
  /**
   * A string which represents a module path that exports a class which maps this database model e.g. './models/some-model'
   */
  classPath?: string;
  /**
   * A string which holds the database object of this model. If this property is missing the database object's name is the concatenation of the model's name and the keyword 'Base' e.g. UserBase, PersonBase etc
   */
  source?: string;
  /**
   * A string which holds the database object that is going to be used for fetching data. If this property is missing this database object's name is the concatenation of the model's name and the keyword 'Data' e.g. UserData, PersonData etc
   */
  view?: string;
  /**
   * A string which represents the version of the model's schema. This version is going to be used in model upgrade operations e.g. 1.0, 0.1.2 etc
   */
  version: string;
  /**
   * A boolean which indicates whether model data will be cached or not. The default value is none -no caching-. A conditional caching allows developers to control caching mechanism while fetching data.
   */
  caching?: "none" | "always" | "conditional";
  fields?: DataFieldSchema[];
  constraints?: DataModelConstraintSchema[];
  eventListeners?: DataModelEventListenerSchema[];
  privileges?: DataModelPrivilegeSchema[];
}